feat(server): add new type of event return by the server and store current logged user.

+++ libs/protocol/src/client/p_client_listen.c
@@ -12,7 +12,7 @@ p_payload_t *p_client_listen(const p_client_t *client)
     p_payload_t *payload = calloc(1, sizeof(p_payload_t));
     const int fd = client->network_data.sockfd;
 
-    if (payload && read(fd, &payload->packet_type, sizeof(uint8_t))
+    if (payload && read(fd, &payload->packet_type, sizeof(uint16_t))
         && read(fd, payload->data, DATA_SIZE))
         return payload;
     free(payload);
